Use configuration file for defining sources instead
Previously, sources were defined as hardcoded values in the application
through ENABLED_SOURCES. This commit transitions the application to use
a configuration file to handle the source configuration so that a user
can define which sources to use.

This commit also removes the support for environment variables and the
.env file. At least currently.

Pyfakefs was introduced also as a dev dependency to support a better way
to test file system based test cases. There may be test cases left which
don't yet use pyfakefs but should.